flux-top: show QUEUE column if queues configured
Problem: Presently the QUEUE column is shown in flux-top if any
jobs have a queue defined.  This is problematic because the QUEUE
column will not be displayed until a job has been submitted, even
when queues are configured.  In addition, the iteration of jobs
is unnecessary.  We can just check if queues are configured in the
Flux config.

Solution: Check if the QUEUE column should be displayed just by
checking if queues are configured in the Flux config.