resource: drain: return EINVAL for excluded ranks
Problem: The resource module returns EEXIST as the errno when a request
to drain an excluded rank is made, but this implies that the node is
already drained, which is misleading.

If the drain RPC fails only due to ranks being already excluded,
return EINVAL instead of EEXIST. Not only is this the more appropriate
errno, but it will also suppress the hint about use of --force in the
`flux-resource drain` error output.