libpmi: use next plugin if libpmi2 fails
Problem: when Flux encounters the Slurm PMI2 library, it
selects the libpmi2 upmi plugin, then fails PMI2_Job_GetId()
outside of a Slurm job (e.g. as a standalone singleton).

Perform PMI2_Init() and PMI2_Job_GetId() in the "preinit" callback
that is used in plugin selection, so that upmi can fall through
to the next plugin if either of those fail.

Fixes #5057