libpmi: use next plugin if libpmi fails
Problem: the libpmi upmi plugin does not attempt to use
a libpmi.so dso before returning success in its "preinit"
callback, so it does not know if it will actually work.

Restructure as was just done in the libpmi2 plugin to call
PMI_Init() and the info functions as part of the preinit
callback that is used in plugin selection.