broker: call sd_notify("STATUS=") on state change
Problem: when the broker is stuck waiting for its TBON
parent to enter RUN state, or for a TBON child to shut down,
this information is hard to diagnose.

Notify systemd each time the broker enters a new state so
information about the current state can be displayed by
'systemctl status flux' in the Status: row.

Skip the states that run rc scripts since the previous commit
already sends notifications about those.