broker: call sd_notify("READY=1")
Problem: systemd needs to know when the broker is "ready"
so it can allow systemctl start to unblock, and can process
dependent units.

For rank 0, send the notification when the broker reachs RUN state.

For rank > 0, send the notification when the broker reaches JOIN state
since this state persists until the upstream node reaches RUN state
and we want to avoid making cluster startup processes deadlock-prone.
(For example, 'pdsh -a systemctl start flux').