broker: add broker.sd_notify attr
Problem: libsystemd is an optional dependency for flux-core
but with Type=notify, it will not be possible for systemd to
start the broker without sd_notify() support.

Add a broker attribute 'broker.sd-notify'.  If set, then the
broker treads missing systemd support as a fatal error.

Set a flag so that sd_notify() calls can be made conditionally.