libsubprocess: remove assert(0) in error path
Problem: In state_change_next(), if an unexpected state is passed
in an assert(0) is called to signify "shouldn't reach here".  This
is bad coding style.

Solution: Instead add an assert at the top of the function that
checks if only the expected states are passed when the function
is called.  Add a note that one state change is not yet handled
due to an issue to corrected in the future.