libsubprocess: avoid assert(0) in API function
Problem: In flux_subprocess_stream_status() an assert(0) is called
if the subprocess is a remote subprocess.  In flux_subprocess_stream_start()
and flux_subprocess_stream_stop(), an EINVAL error is returned on remote
processes.

Make flux_subprocess_stream_status() consistent to flux_subprocess_stream_start()
and flux_subprocess_stream_stop() and just return EINVAL if the function
is called on a remote subprocess.