broker: simplify module loading
Problem: broker module management is overly complex.

When loading a module, assume that the specified name can be
directly translated to a file path, either by adding ".so" and
trying to open it in each directory of the FLUX_MODULE_PATH, or
by opening it directly.

Allow a module to be unloaded by path as well.