broker: make mod_name symbol optional
Problem: the broker module's 'mod_name' is used internally
to define its name, but the module loading infrastructure
now uses the DSO file name to derive the name.

Let 'mod_name' be optional and derive the name from the file name.
If 'mod_name' is defined, it must be the same as the file name.

Rename job-manager-dummy.so (part of t2100-job-ingest.t) that
violated this rule.