job-exec: improve job shell/imp execution error
Problem: When the configured job shell or IMP fails to exec,
e.g. because of a missing executable or incorrect path, the error
raised by the job-exec module is confusing, e.g.:

  job shell exec error on broker <host> (rank <id>): No such file or directory

Since this error message is included as the job exception note, it
should be made clearer.

Improve this message to include the path to the failing
executable. This will allow users and admins to easily determine what
job shell or IMP path is failing.