flux-job: attach: fix hang for jobs without start event
Problem: If a job never starts, i.e. a start event is never posted
because there was a problem executing the job shells or IMP process,
then `flux job attach` could hang because the notify timer may still
be holding a reference to the reactor.

Be sure to stop the notify_timer in attach_completed_check() so it no
longer keeps the reactor active when attach thinks the job is complete.