testsuite: job-exec: add test for missing job shell
Problem: No test in the testsuite ensures that an execution error
for the configured job is handled as expected by the job-exec module.

Add a test to t2400-job-exec-test.t which ensures that a missing job
shell results in 127 exit code from `flux run` and that the path to
the job shell is part of the emitted error.

Skip the test on CentOS 7 since this test fails on that distro for
an as yet undiagnosed reason. Instead of returning an exec error
when launching the job shell from a bad path, success for rexec.exec
is returned to job-exec, and the process then later exits with code
127. So instead of raising a fatal job exception, the job just exits
with failure on this distro.