broker: set module name in flux_t aux container
Problem: a module that can be loaded under multiple names
needs to be able to determine what the name is so it can register
approriate message handlers.

Since we are already setting the module UUID in the module flux_t
handle's aux container (flux::uuid), do the same for the name (flux::name).