libflux: add FLUX_DEPRECATED macro
Problem: In the future we may wish to deprecate libflux functions
by adding the deprecated attribute to function definitions.  We also
need a way to ensure the python bindings do not generate bindings
for those deprecated functions.

Add a FLUX_DEPRECATED macro that will add "__attribute__((deprecated))"
to functions the macro wraps.  The macro will be empty when python
bindings are generated, ensuring the function will not be generated
for python.