build: don't export 'mod_service' in broker modules
Problem: the 'mod_service' symbol is in the default broker
module symbol map, but that symbol is not used.

Drop mod_service from the default module map.
Drop mod_service and mod_name from the custom job manager module map.

mod_name is still supported in out of tree modules, so although it is
not used in any of the in tree modules, it remains in the map for testing
purposes.