libjob/test: ignore deprecated test warnings
Problem: flux_job_list() and flux_job_list_inactive() are now deprecated,
so we see warnings when compiling unit tests for libjob.  We don't want
to remove the tests, because the tests still cover the deprecated
functions.

Solutions: Add -Wno-deprecated-declarations to the CFLAGS for job unit
tests to avoid deprecated-declaration warnings.