testsuite: ensure flux python is first in PATH
Problem: If flux is configured with a non-default python3, then some
test scripts that use `#!/usr/bin/env python3` may fail since the
default python3 in PATH is not the version configured with flux.

Prepend the path to the `flux python` executable in PATH so it is found
before any other python3. Then prepend $FLUX_BUILD_DIR/src/cmd to PATH
so that the built flux(1) executable is found before any flux that shares
a path with `flux python`.