libsubprocess: avoid assertion failure in remote_output()
Problem: A broker crash was recently observed in libsubprcess
remote_output() from an assertion failure in zhash_lookup() due
to a NULL target hash.

Diagnosis of the underlying cause is still in progress, but for now
avoid the abort by checking that the subprocess has a valid channels
hash (and that stream is not NULL) before going on to process the
output. Log an error if the condition is detected.