flux-top: notify user when queue has no jobs
Problem: If a queue has no jobs to display, it can confuse a user
why nothing is showing up in flux-top, especially when flipping through
queues with the left and right keys.

If there are no jobs to display for a specific queue, write a message
in the joblist pane indicating that is the case.