libsubprocess: avoid segfault on error path
Problem: libsubprocess can trigger a segfault if a remote
subprocess object is destroyed before the remote protocol
is complete.

The remote's future should be held by subprocess->f but in the
code cleanup of 8bb3e74fad8c2178a8900121cf7aa1ef69869766, this
was dropped and it now has an independent life cycle from the
subprocess.

Assign the future to subprocess->f and don't destroy it separately
on completion in the continuation.

Fixes #5094