build: add `make deb` target to build test debian package
Problem: There is no provided method to package a build directory
of flux-core for convenient installation for testing.

Add a set of debian packaging metadata under the debian directory
and a script in scripts/debbuild.sh which triggers a packaging build
of the current source.

For convenience, a `make deb` target is also added to the top-level
Makefile.am, which uses the new script to build a debian package.

By default DEB_BUILD_OPTIONS is set to "nocheck terse", such that
no `make check` is run during the package build. This environment
variable should be able to be overridden on the command line.