testsuite: fix setup error in system tests
Problem: In t9000-system.t the FLUX_TEST_INSTALLED_PATH variable
is setup if 'debug' is set.  However, the check for 'debug' is
done before sharness is sourced, so the `-d` option when running
the tests doesn't matter.  This lead to the t9000-system.t tests
failing when running make check in a docker container.

Solution: Instead of checking for 'debug' to set FLUX_TEST_INSTALLED_PATH,
check for FLUX_ENABLE_SYSTEM_TESTS.