.typos.toml: add false positive typos config
Problem: The current typos workflow check only checks a select
number of files because there are so many false positives in
flux-core.

Solution: Add a .typos.toml configuration file to eliminate false
positives so we can run the typo check on all of flux-core.  The
config file either skips files with lots of false positives or
lists common false positives we wish to ignore.

Fixes #5075