libschedutil: skip service unregistration
Problem: when the scheduler fails to register the "sched"
service name, it attempts to unregister the service twice
on the error path.

Since dynamically registered service names are automatically
unregistered when a module is unloaded, we don't even need
this code.

Drop it and simplify sched_ops_register() and _unregister().