job-exec: Use FLUX_IMP_EXEC_HELPER when possible
Problem: The flux-security IMP now supports a helper process for
fetching its input instead of taking input on stdin, and there is a
flux-imp-exec-helper command provided in flux-core, but the job-exec
service doesn't yet support setting up the IMP to use the helper.

Add an exec config function `config_use_imp_helper()` which defaults
to true if flux-security v0.9.0 or above is detected (This is the
first version that supports an IMP helper command). When the IMP
helper is configured, avoid unnecessarily fetching J for multi-user
jobs in the job-exec module, and instead set FLUX_IMP_EXEC_HELPER to
"flux imp-exec-helper JOBID". The IMP will then use this command to
fetch its input instead of reading from stdin.

Support no-imp-helper module option to the job-exec module to
allow disabling use of the helper even with flux-security v0.9.0.