shell: mpir: fix encoding of hosts with leading zeros in prefix
Problem: The mpir 'nodelist' implementation stores the entire
numerical suffix as a rangelist object, but the conversion to/from
integers drops all leading zeros if present in the hostname suffix.
E.g. node0212 will end up stored as prefix "node", suffix 212, so
the hostname is converted from node0212 to node212.

Advance past leading zeros when splitting the hostname, such that
leading zeros are in the hostname prefix instead of the suffix.
This ensures that the suffix is always representable as an integer
without loss of characters.

Fixes #5115