shell: mpir: document mpir JSON encoding in source
Problem: The mpir rangelist, nodelist, and proctable encodings are
not documented at all, which makes it confusing when there are bugs
or issues.

Add a documentation header to each of the aforementioned source
files briefly describing the encoding used for each of these JSON
representations.

This should at least help future developers reason about what the
JSON input and output from these functions is representing.