python: importer: return empty list only if no plugins found
Problem: The flux.importer.import_plugins() function returns an
empty list if there is a ModuleNotFoundError from any plugin in the
namespace being imported. This is because the import_plugins_pkg()
call was placed inside the try block, so errors from that function
also result in an empty list.

Exceptions raised in import_plugins_pkg() should be handled by the
caller, not locally, so move this function out of the try/except block.

We still check for ModuleNotFoundError from importlib.import_module()
since that exception will be raised for an empty directory, and we
want to reaturn an empty list in that case.

Fixes #5116