doc: document the {{name}} tag in output file templates
Problem: The use of the {{name}} tag in job output and error filename
templates is not documented.

Document the {{name}} tag in man1/common/submit-standard-io.rst so
it appears in the appropriate manual pages. Now that there is more
than one supported tag, use a definition list instead of defining
supported tags inline.