testsuite: ensure user dbus is running
Problem: sdbus tests only check that the systemd user instance
is running, not that sdbus is.

The systemctl program communicates with systemd over a private
socket and does not ensure that the user dbus daemon is also
running.

Add a call to "busctl --user status" and skip the test if it fails.

Fixes #5127