libutil: fix underflow error in timestamp_parse()
Problem: There is a potential for underflow in the calculation
of the optional microseconds in the timestamp string provided to
timestamp_parse(). This occurs when the double precision result
of strtod(3) is multiplied by 1e6 to get the microseconds as an
integer. The integer truncates the floating point part so, for example,
1234.9999 wil be truncated to 1234 when it should be 1235.

Add 0.5 to the result of multiplying the double value by 1e6 so that
truncation returns the correct value.