job-list: add state matching library
Problem: In job-list we store jobs on three lists, pending,
running, and inactive.  If we begin to use RFC31 constraints
to filter jobs, it would be inefficient to scan all three lists
for every job list query.

Solution: Add a job state matching library.  It will determine if
jobs in the pending, running, or inactive state are possible given
the job constraint sent with the job-list query.  For example:

states=pending AND userid=42

In the above constraint, a job in a pending state can possibly be
matched to this constraint, so it is worthwhile to scan the pending
job list for jobs with userid=42.  However, it is impossible for
a job in running or inactive state to match at all, so they should
not be scanned.

This will allow job-list queries to run more efficiently if entire
lists can be skipped if it is impossible for any jobs on those lists
to be matched to a constraint.

Add unit tests.