libflux: return remote parent-uri if FLUX_PROXY_REMOTE set
Problem: The parent-uri broker attribute is always encoded as a local
URI, since currently the parent broker always shares a node with
its children. This presents a problem when connecting to a remote
instance, since users of the parent-uri attribute will attempt to
connect to a local socket which does not exist.

Allow the parent-uri attribute to be transparently aliased to a remote
URI in attr_get() if the environment variable FLUX_PROXY_REMOTE
is set. If set, this environment variable should contain the full
"authority" part of the URI that flux-proxy(1) used to connect to the
instance. The parent-uri is then converted to a remote URI using the
contents of this variable.