broker: add FLUX_PROXY_REMOTE to runat env blocklist
Problem: flux-proxy(1) may set the FLUX_PROXY_REMOTE environment
variable to influence handling of the parent-uri broker attribute,
but this variable only makes sense for processes that are direct
descendants of the flux-proxy(1) process.

Add this variable to the broker runat env blocklist. This ensures
that a broker started locally, e.g. via `flux start --test-size`
doesn't have this environment set in the initial program or other
processes spawned as a child of the broker.