flux-proxy: export FLUX_PROXY_REMOTE
Problem: When flux-proxy(1) connects to a remote instance, the broker
parent-uri attribute is encoded as a local URI that is not valid in
the proxy environment.

If flux-proxy(1) has connected to a remote URI, export the 'authority'
part (username, host, etc.) in the FLUX_PROXY_REMOTE environment
variable.  This gives a hint to flux_attr_get(3) and derived calls
to transform any parent-uri to a remote URI, which should make all
commands and interfaces that use this URI (e.g flux_job_timeleft(3),
flux-pstree(1), etc.) transparently work from the remote proxy session.

Fixes #5132