libflux: plugin: improve error on flux_plugin_init failure
Problem: When a plugin is loaded by flux_plugin_load_dso() and its
flux_plugin_init() callback fails, no text error message is set on
the plugin which results in an empty error message.

Add an error message that indicates that the plugin init callback
failed so there is a hint as to what went wrong.