libflux/test: test failure of flux_plugin_init()
Problem: No unit tests of the libflux plugin code ensure proper
handling of a failure of a plugins flux_plugin_init() callback.

Add the ability to trigger a failure in one of the test plugins,
and use this to ensure a nonzero return from flux_plugin_init()
causes a failure in flux_plugin_load_dso() and that an appropriate
error message is returned from flux_plugin_strerror().