broker: make tbon.connect_timeout configurable
Problem: there is no way to tune the ZMQ_CONNECT_TIMEOUT
socket option.

Allow the connect timeout to be set either in the config:
  [tbon]
  connect_timeout = "10s"

or on the broker command line:
  -Stbon.connect_timeout=10s

The broker command line takes precedence.

Set a default of 30s instead of the system default of "a long time"
per zmq_setsockopt(3).

Fixes #5134