shell: unset FLUX_PROXY_REMOTE for job tasks
Problem: FLUX_PROXY_REMOTE may be set in the environment of a job
that is submitted under flux-proxy(1), but this environment variable
doesn't make sense for job tasks since they are always local to a
broker at this time.

Unset FLUX_PROXY_REMOTE for all job tasks in the job shell.