doc: clarify --since in flux-jobs(1)
Problem: The wording in the --since option in flux-jobs(1) is a bit
misleading, suggesting that the --since option may apply only to
completed/inactive jobs.

Solution: Note that the --since option applies to jobs that were
active after the given timestamp, which makes it clear that --since
also will list pending and running jobs.