libioencode: allow NULL data parameter in iodecode()
Problem: There is no way to query the decoded length of data using
iodecode() without unpacking the data itself, resulting in unnecessary
allocation and memory copying.

Allow data to be NULL when len is not. In this case unpacking the data
is skipped, but the length is still assigned to to the lenp parameter.