shell: truncate KVS output above a threshold of 10MB
Problem: A job that writes a huge amount of data to the KVS can
wreak havoc when the output eventlog is read back since the KVS has
to reassemble the eventlog into one big value to return to the user
or job-info service.

Limit the maximum amount of data a job can write to the KVS in the
shell output plugin to add a first-level protection against this issue.
An arbitrary limit of 10MB is chosen, after which output for the given
stream is trunctated. A warning is issued at the time of truncation
and also at the end of the job, e.g.

  flux-shell[0]:  WARN: output: stdout will be truncated, 10MB limit exceeded
  flux-shell[0]:  WARN: output: stdout: 5514240 bytes truncated