broker: forward non-deadly signals to all jobs
Problem: The broker blocks all signals in main(), except those for
which a core dump may be generated, then handles a specific smaller
set of "fatal" signals in broker_handle_signals(). This leaves all
other signals blocked, and any attempt to send one of these signals
to a batch job will be ignored since the job shell delivers them
directly to the broker.

Add SIGUSR1 and SIGUSR2 to the list of signals handled in
broker_handle_signals(), and instead of calling runat_abort() for
these 2 signals, forward them to all jobs via the job-manager.killall
RPC. If the RPC fails, send the signal locally so the signal goes
somewhere.

For now, keep the current behavior for the set of fatal signals that
terminate rc2 to avoid breaking anything that depends on this behavior.