testsuite: update dummy job shell for barrier on stdin/out
Problem: The job-exec module may transition to using stdin/out for
the job shell exec protocol, but the dummy job shell does not yet
support this.

If FLUX_EXEC_PROTOCOL_FD is not set, default to using stdin for
the read side of the protocol, and stdout for the write side. Also,
ensure that the dummy job shell does not emit other output to stdout -
instead use stderr for all logging.