testsuite: FLUX_EXEC_PROTOCOL_FD is no longer required
Problem: There is a test in t2602-job-shell.t that ensures the shell
fails if FLUX_EXEC_PROTOCOL_FD is unset, but soon the absence of this
environment variable will just mean that stdin/out are used for the
protocol, so this should not be an error.

Remove the invalid test.

Also, export a bad value for FLUX_EXEC_PROTOCOL_FD to ensure this
environment variable is passed down through the test execution.