shell: use stdin/out if FLUX_EXEC_PROTOCOL_FD not set
Problem: The job shell assumes FLUX_EXEC_PROTOCOL_FD will be set,
but in the future job-exec will use stdin/out for the exec protocol
and will not set FLUX_EXEC_PROTOCOL_FD.

If FLUX_EXEC_PROTOCOL_FD is not set, default to using stdin/out for
the exec protocol in the job shell. Split shell->protocol_fd into
an array of 2 fds to support separate read and write fds.