job-exec: use stdin/out for exec proto when using imp helper
Problem: When the job-exec module uses the imp helper script, this
frees up the job shell's standard input for use by the exec protocol.
This may be necessary when invoking the shell under systemd, where
creation of a separate socketpair for FLUX_EXEC_PROTOCOL_FD may not
be feasiable.

When using the imp helper do not set FLUX_EXEC_PROTOCOL_FD and use
stdin/out for the job-exec protocol.