flux-jobs: correct behavior with -a and --filter
Problem: If the user specifies both the -a and --filter options, all jobs
are currently output (i.e. -a takes precedence).  This can be unexpected
because the --filter should seemingly "filter" whatever is otherwise
output.

If the user specifies both options, do not use the -a option and output
a warning to the user.

Fixes #5142