sharness: add function to check for COLUMNS usability
Problem: With some versions of bash (likely >= 5.0), the COLUMNS shell
variable, when overridden in the environment for testing purposes, is
not preserved across invocation of the flux(1) command driver during
sharness tests. This appears to occur because COLUMNS is reset by
the libtool wrapper script (which has the shebang line #!/bin/bash),
and only seems to occur if the sharness test is run with `-d -v`
for unknown reason.

Add a function to sharness.d/flux-sharness.sh that checks if COLUMNS
can be preserved in the testsuite environment across an invocation of
`flux(1)`. This can be used by tests that require this functionality
to set a prereq, allowing them to skip the tests if necessary instead
of erroneously failing.