testsuite: skip tests that use COLUMNS if value not preserved
Problem: A couple tests in the testsuite explicitly set the COLUMNS
environment variable to ensure the output of a command responds to
changes in the terminal width. However, in some environments the
flux(1) libtool wrapper script can reset this variable, causing the
tests to erroneously fail.

Use the provided test_columns_variable_preserved() function to check
for the required behavior of this variable, and skip the tests if not.