python: format all zero UtilDatetime objects as an empty string
Problem: It is inconvenient that UtilDatetime objects (obtained via the
`!d` conversion specifier) are treated differently than the normal
datetime conversion (`!D`) when the timestamp is zero. In the first
case, the datetime is rendered to the epoch date in whatever format
is specified by the user, while in the second case the result is an
empty string. Since a zero timestamp is considered "unset", the empty
string is much more preferable here.

Since the prime use case for UtilDatetime is in output format rendering
for utilities, have this class return an empty string by default for
zero timestamps, presuming these are unset.

Fix one test in t2800-jobs-cmd.t that assumed a 00:00 result.