python: support `h` suffix with `!d` conversion specifier
Problem: The `!d` conversion specifier converts timestamps to
UtilDatetime objects which support a `::` separator in order to allow
a format after the datetime format specification. However, since the
format is applied within the UtilDatetime format() method, the `h`
suffix is not supported to replace empty fields with a hyphen.

Check for a `h` suffix in the UtilDatetime.format() method and replace
an empty string with `-`.

Ignore UtilDatetime objects where `h` is normally handled so that
the suffix can later be processed in UtilDatetime.format()