python: add default format for `!d` conversion specifier
Problem: The `!d` conversion specifier returns an empty string if
no format is provided, which could cause confusion in casual usage.
For example `{expiration!d}` returns an empty string even if there
is an expiration timestamp for a job.

If no format is provided, default to an ISO 8601 format "%FT%T".