libutil: add base256, a binary to emoji encoding
Problem: There is a distict lack of an encoding that represents binary
data as emoji.

Adapt the code from:

 https://github.com/Equim-chan/libb256

which implements a "base256" encoding from binary to emoji using two
tables of emoji to translate each byte of input into a single emoji.

Update the code for better usage within flux-core. Drop the internal
hash for decode - we don't care if decode is slow at this point and
implementing the hash safely would be tricky.

Implement input checking and errno returns instead of specialized
return codes for errors.

Add a prefix character which can be checked on decode, and a
is_base256() macro to check if a string is encoded in base256.