python: support W presentation type
Problem: When outputting wide characters (e.g. emojis) in flux-jobs,
the alignment of output can be poor due to the characters having
different output widths.

Solution: Add a new W presentation type that can adjust formating
of the form "(<|>)N", e.g. {id.emoji:>12W}.  The output width will
be adjusted given the number of wide characters that exist in the
string.

Note that this presentation type does not help for all output scenarios,
as it depends on the width of the output and width of the alignment,
but it definitely helps in some scenarios.