flux-job: attach: ignore stdin if read watcher returns EINVAL
Problem: Users have reported an error from `flux job attach`:

 flux-job: flux_buffer_read_watcher_create: Invalid argument

The underlying cause of which is not well understood. However, it
seems unnecessary to terminate flux-job attach due to this error. In
most cases, the job is not processing stdin.

Turn the fatal error when `flux_buffer_read_watcher_create()` fails
into a warning.