flux-job: submit: strip whitespace from pre-signed jobspec
Problem: Leading and trailing whitespace around signed jobspec, or J,
is preserved when submitting with `flux job submit --flags=signed`,
but the whitespace is not part of a sign-wrapped jobspec, so it would
be best ignored.

If the provided file to `flux job submit` is already signed, strip
leading and trailing whitespace before passing to `flux_job_submit(3)`.

Update one test in t2404-job-exec-multiuser.t which was required to
use the jq(1) `-j` option to extract J from the IMP's input to avoid
adding an extra newline (J was being submitted with one already
present). Now that J is guaranteed to not have extra whitepsace,
we can safely revert the test to the more typical `jq -r`.