job-exec: always use stdio for exec barrier
Problem: job-exec uses either stdio or a subprocess channel
to implement the exec barrier, but now that we require a
minimum flux-security version of 0.9.0, stdio may be used in
all cases.

The critical change in flux-security is that the IMP can now
use a helper program to obtain J in lieu of reading it from stdin.

Simplify the code in job-exec so that stdio is always used
for the exec barrier and the IMP always uses the helper.

This also makes it easier to integrate pending work to execute jobs
as systemd transient units, since subprocess channels will not work
in that environment.