broker: always execute rc2 in a separate process group
Problem: The broker may execute rc2 (the intiial program) using
either runat_push_shell_command() (if there is only 1 argument),
or runat_push_command() (if there are multiple arguments). However,
runat_push_shell_command() runs processes in their own process group,
while runat_push_command() does not. This causes different behavior in
signal handling between different batch jobs depending on the number
of arguments: i.e. in one case, signals sent to the batch job are
delivered to the batch script, and in the other case they are not.
This difference makes the behavior of signals unpredictiable for user
batch jobs.

Add the FLUX_SUBPROCESS_FLAGS_SETPGRP flag in runat_push_command()
as it is done in runat_push_shell_command() so that rc2 is always run
in its own process group, regardless of the number of args.