testsuite: fix race in flux top tests
Problem: On rare occasions some tests in t2801-top-cmd.t fail
when running large parallel runs of the testsuite.  The suspicion
is that the job-list module has not yet synced up to have knowledge
of the jobs that have been submitted in the tests.

Add a function to wait until jobs in t2801-top-cmd.t are known
by job-list and have reached the state we want to ensure tests
will pass.