testsuite: update python type check
Problem: In the set_and_check_context() helper function, the test
checks that the returned type is not 'bytes'.  This is a non-optimal
check as it doesn't check for the type we are expecting.  In addition,
a future test for a 'bytes' type can't be done.

Solution: Have set_and_check_context() take an expected type as a
parameter.  Ensure that the returned type is the type we are expecting.