testsuite: test Python KVS more thoroughly
Problem: In python/t0005-kvs.py, the set_and_check_context() function
uses the same KVSDir object to write and read a value back.  This doesn't
ensure that the KVS data was actually commited to the KVS successfully,
as the data may just be locally cached in the KVSDir object.

Solution: After performing a KVS commit, open a new KVSDir object to ensure
the data has been written to the KVS correctly.