libtuil: add signal name/number helper routines
Problem: Conversion of signal name to number and vice versa will be
useful in a couple places in flux-core, but there are no standard
routines for this purpose.

Add a sigutil interface with two helper routines:

 int sigutil_signum (const char *s);

which converts a string to a signal number, and:

 const char * sigutil_signame (int signum);

which converts a signal number to a name.

The set of supported signals is limited to the most common via a
statically defined list in sigutil.c.