job-list: allow list-id to wait for job state
Problem: It'd be convenient if the job-list.list-id service not only
waited for a job id to be legal but to also waited for the job to
pass a certain job state.

Solution: Support an optional job state in the job-list.list-id service.
Return when a job passes this job state or reaches the inactive state.

Fixes #2864