shell: add plugin to optionally signal job before expiration
Problem: There is no option to signal a job some grace period before
the job expiration.

Add a 'signal' shell plugin that checks for a "signal" shell option
with the contents:

 { "signal"?i "timeleft"?i }

If present the signal plugin will set up a periodic watcher to trigger
timeleft seconds before any job expiration. The callback will send
the defined signal (SIGUSR1 by default) to the job. By default the
value for timeleft is 60s and the default signal is SIGUSR1.