python: add --signal=sig@time option to submission commands
Problem: There is not a simple interface to use the job shell signal
option from the command line of the submission commands.

Add a --signal=[SIG][@TIME] option to the base submission command
option set. This option allows users to easily specify a signal to send
to the job along with a time remaining at which to send the signal.

Either part of the option argument may be omitted, so

 --signal=USR1

or

 --signal=@30s

will work as expected. The default signal is SIGUSR1 and time is 60s.