job-list: store jobid in idsync list struct
Problem: In the idsync API a list is stored in a hash and
an externally stored jobid in a list entry is used as the key.
This currently is fine because the entire list is read back and
destroyed in idsync_check_waiting_id().

However, if we ever want to read/delete some list entries but not
others, the externally stored jobid makes the hash lookup unsafe,
as the storage of the key can no longer be guaranteed.

Solution: Store the list in a new struct that also stores the
jobid.  Also, adjust the name of a cleanup function as a result.