python: OutputFormat: run pre() only after line is printed
Problem: The `pre` argument to OutputFormat.print_items() is
documented as being called before printing each line, but it is
called even if an empty line is generated. This can cause problems
in tools like flux-jobs(1) which uses a pre() function to colorize
lines of output. If a color is set, but the line is  then skipped
or an exception is thrown, the terminal may be left with the line
color enabled.

Fix OutputFormat.print_items() to call the pre() function after the
line has been formatted, just before printing it. This removes the
possibility that pre() will be called without post() due to an empty
line or an exception during line formatting.

Fixes #5209