python: add filter method to ResourceStatus class
Problem: There is a need to support filtering the result of the
flux.resource.resource_status() function used by the flux-resource(1)
`status` and `drain` commands.

Add a `filter` method to the ResourceStatus class which can filter the
various idsets available in this class based on an idset of 'include'
ranks. This method can then be used to support an `--include` option in
the flux-resource(1) command.